Create MXDChecker.py
creates a csv report of all MXDs in a directory. The report will list the layers in the MXD and print any errors the tool had in an attempt to open the MXD. Caution should be used if you have a lot of files! If the MXD is corrupt the tool is designed to stop attempting to open it and move to the next one. I need to update this as it is currently hard-coded.